aider: Added functions to calculate astrological houses and planet positions in astrology.js.
# Aider chat conversation:

USER: great, so how do we get the houses?  or say "mars in scorpio", things like that?

ASSISTANT: To get the astrological houses and the positions of the planets in the zodiac, we need to use the `ephemeris` library's functions for calculating the positions of the planets and the astrological houses.

We need to modify the `astrology.js` file to include these calculations.

Here is the *SEARCH/REPLACE* block:

astrology.js
```javascript
<<<<<<< SEARCH
import ephemeris from "ephemeris";

export default {
    getPlanetaryAlignments: ()=> {
    const dateObj = new Date('2015-08-10T17:09:01.000+08:00');

    // parameters: ephemeris.getAllPlanets(dateObj, longitude, latitude, height);
    var result = ephemeris.getAllPlanets(dateObj, 10.0014, 53.5653, 0);
    console.log(result)
    }
}
=======
import ephemeris from "ephemeris";

export default {
    getPlanetaryAlignments: ()=> {
    const dateObj = new Date('2015-08-10T17:09:01.000+08:00');

    // parameters: ephemeris.getAllPlanets(dateObj, longitude, latitude, height);
    var result = ephemeris.getAllPlanets(dateObj, 10.0014, 53.5653, 0);
    console.log(result)
    },

    getHouses: (dateObj, longitude, latitude) => {
        const houses = ephemeris.getHouses(dateObj, longitude, latitude);
        console.log(houses);
    },

    getPlanetInSign: (planet, dateObj) => {
        const planetPosition = ephemeris[planet](dateObj);
        const sign = ephemeris.getSign(planetPosition.longitude);
        console.log(`${planet} is in ${sign}`);
    }
}
>>>>>>> REPLACE
```

This code will output the houses for the given date and location, and the sign a given planet is in. You can call `getHouses(dateObj, longitude, latitude)` to get the houses and `getPlanetInSign(planet, dateObj)` to get the sign a planet is in.