aider: Use the `apparentLongitudeDd` property to get the longitude of the planet in the `getPlanetInSign` function.
# Aider chat conversation:

USER: planetPosition.longitude is not the right variable.  which one do we use?  this is the planetPosition object: planetPosition {
  name: 'sun',
  raw: {
    weight: 100,
    key: 'sun',
    position: {
      equinoxEclipticLonLat: [Object],
      date: [Object],
      lightTime: 8.430979278000157,
      aberration: [Object],
      constellation: 22,
      apparent: [Object],
      apparentLongitude: 137.44127998638336,
      apparentLongitudeString: `137°26'28"`,
      apparentLongitude30String: `17°26'28"`,
      geocentricDistance: -1,
      altaz: [Object]
    }
  },
  apparentLongitudeDms30: `17°26'28"`,
  apparentLongitudeDms360: `137°26'28"`,
  apparentLongitudeDd: 137.44127998638336,
  geocentricDistanceKm: -1
}

ASSISTANT: From the `planetPosition` object, it seems like the `apparentLongitudeDd` property is the one that should be used to get the longitude of the planet. This property represents the apparent longitude of the planet in decimal degrees.

To fix this, we need to modify the `getPlanetInSign` function in the `astrology.js` file.

Here is the *SEARCH/REPLACE* block:

astrology.js
```javascript
<<<<<<< SEARCH
        const sign = ephemeris.getSign(planetPosition.longitude);
=======
        const sign = ephemeris.getSign(planetPosition.apparentLongitudeDd);
>>>>>>> REPLACE
```
This change will use the `apparentLongitudeDd` property of the `planetPosition` object to get the sign of the planet.